backend/requirements.txt
# FastAPI Backend for AI Video Interview Platform

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.24.1
websockets==12.0
aiofiles==23.2.1

# Database
supabase==2.0.2
postgrest==0.13.0
gotrue==1.3.0
realtime==1.0.0
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0

# Audio Processing
pydub==0.25.1
ffmpeg-python==0.2.0
scipy==1.11.4
numpy==1.26.2

# AI & Speech
openai==1.3.7
anthropic==0.7.8
deepgram-sdk==3.0.0
whisper==1.1.10

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing & Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
isort==5.12.0